# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/certificates": {
            "get": {
                "description": "证书列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "证书"
                ],
                "summary": "证书列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "证书",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "是否包含total",
                        "name": "include_total",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "pre_page",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "展开数据",
                        "name": "expand",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.ListCertificateResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HttpError"
                        }
                    }
                }
            },
            "post": {
                "description": "创建证书",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "证书"
                ],
                "summary": "创建证书",
                "parameters": [
                    {
                        "description": "数据",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.CreateCertificateParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Certificate"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HttpError"
                        }
                    }
                }
            }
        },
        "/certificates/{id}": {
            "get": {
                "description": "获取证书",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "证书"
                ],
                "summary": "获取证书",
                "parameters": [
                    {
                        "type": "string",
                        "description": "证书ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "展开数据",
                        "name": "expand",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Certificate"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HttpError"
                        }
                    }
                }
            },
            "post": {
                "description": "更新证书",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "证书"
                ],
                "summary": "更新证书",
                "parameters": [
                    {
                        "type": "string",
                        "description": "证书ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "数据",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.UpdateCertificateParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Authorize"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HttpError"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除证书",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "证书"
                ],
                "summary": "删除证书",
                "parameters": [
                    {
                        "type": "string",
                        "description": "证书ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HttpError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.Authorize": {
            "type": "object",
            "properties": {
                "attribute": {
                    "description": "鉴权属性",
                    "allOf": [
                        {
                            "$ref": "#/definitions/value.AuthorizeAttribute"
                        }
                    ]
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "description": "描述",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "description": "鉴权备注",
                    "type": "string"
                },
                "type": {
                    "description": "鉴权类型",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dto.Certificate": {
            "type": "object",
            "properties": {
                "certificate": {
                    "description": "证书",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "dns_names": {
                    "description": "证书支持的域名",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "key": {
                    "description": "私钥",
                    "type": "string"
                },
                "name": {
                    "description": "证书备注名",
                    "type": "string"
                },
                "not_after": {
                    "description": "证书有效期",
                    "type": "string"
                },
                "not_before": {
                    "description": "证书开启时间",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "httputil.HttpError": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/httputil.HttpErrorDetail"
                }
            }
        },
        "httputil.HttpErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "service.CreateCertificateParam": {
            "type": "object",
            "required": [
                "certificate",
                "key",
                "name"
            ],
            "properties": {
                "certificate": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "service.ListCertificateResult": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.Certificate"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "service.UpdateCertificateParam": {
            "type": "object",
            "required": [
                "certificate",
                "id",
                "key",
                "name"
            ],
            "properties": {
                "certificate": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "value.AuthorizeAttribute": {
            "type": "object",
            "properties": {
                "binary": {
                    "$ref": "#/definitions/value.AuthorizeAttributeBinary"
                }
            }
        },
        "value.AuthorizeAttributeBinary": {
            "type": "object",
            "required": [
                "sources"
            ],
            "properties": {
                "header": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "sources": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/value.AuthorizeSource"
                    }
                }
            }
        },
        "value.AuthorizeSource": {
            "type": "object",
            "required": [
                "name",
                "source"
            ],
            "properties": {
                "name": {
                    "description": "匹配值",
                    "type": "string"
                },
                "source": {
                    "description": "匹配源",
                    "type": "string"
                }
            }
        },
...
    }
}